requirements.txt
Django==4.2.15
djangorestframework==3.15.2
djangorestframework-gis==1.0
django-filter==24.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
django-storages==1.14.4
boto3==1.34.162
razorpay==1.4.2
Pillow==10.4.0
python-dotenv==1.0.1
drf-spectacular==0.27.2
celery==5.4.0
redis==5.0.7
PyJWT==2.9.0
reportlab==4.2.2
